+++ R/list_groups.R
@@ -1,26 +1,28 @@
 #' List Groups
 #'
-#' @param filter Named Vector. Required. Only One of the two: id or mine.
-#' id ``The Named parameter specifies a comma-separated list of the YouTube group ID(s) for the resource(s) that are being retrieved.''
-#' mine String. Can be 'True' or 'False.' ``Set this parameter's value to true to retrieve all groups owned by the authenticated user.''
-#' @param page_token String. Optional. ``Identifies a specific page in the result set that should be returned.''
+#' @param filter Named Vector. Required. Only one of the two: id or mine.
+#' id: Comma-separated list of YouTube group ID(s) to retrieve.
+#' mine: Set to TRUE to retrieve all groups owned by the authenticated user.
+#' @param page_token String. Optional. Identifies a specific page in the result set that should be returned.
 #' @param \dots Additional arguments passed to \code{\link[tubern]{tubern_GET}}.
 #'
 #' @return named list
 #'
 #' @export
 #'
-#' @references \url{https://developers.google.com/youtube/analytics/v1/reference/groups/list}
+#' @references \url{https://developers.google.com/youtube/analytics/reference/groups/list}
 #'
 #' @examples
 #' \dontrun{
-#' list_groups(filter = c(mine = 'True'))
+#' list_groups(filter = c(mine = TRUE))
 #' }
 
-list_groups <- function (filter, page_token, ...) {
+list_groups <- function (filter, page_token = NULL, ...) {
 
-  querylist <- list()
-  querylist <- c(querylist, filter)
+  querylist <- as.list(filter)
+  if (!is.null(page_token)) {
+    querylist$pageToken <- page_token
+  }
 
   res <- tubern_GET("groups", querylist, ...)